2.3.0.1 - Mutations, NPC Queues, FFA Event, Timelimit, etc

+++ src/me/santipingui58/splindux/npc/SplinduxNPC.java
@@ -21,27 +21,31 @@ public class SplinduxNPC {
 	public SplinduxNPC(NPC npc,NPCType type) {
 		this.type = type;
 		this.npc = npc;
-		double x = npc.getStoredLocation().getBlockX()+0.5;				
-		double z = npc.getStoredLocation().getBlockZ()+0.5;
-		
+		 createHologram();	
+	}
 	
+	
+	public void createHologram() {
+		double x = npc.getStoredLocation().getBlockX()+0.5;	
+		double y =  npc.getStoredLocation().getBlockY()+3;
+		double z = npc.getStoredLocation().getBlockZ()+0.5;	
+		if (this.type.equals(NPCType.FFA)) {
+			 y = npc.getStoredLocation().getBlockY()+2.75;	
+		} 
+		Location l = new Location (npc.getStoredLocation().getWorld(),x,y,z);
+		this.hologram = loadHologram(l);
+		hologram.clearLines();
+		
 		if (this.type.getGameType()!=null) {
 		if (this.type.equals(NPCType.FFA)) {		
-			double y = npc.getStoredLocation().getBlockY()+2.75;
-			Location l = new Location (npc.getStoredLocation().getWorld(),x,y,z);
-			this.hologram = HologramsAPI.createHologram(Main.get(),l);
-			hologram.clearLines();
 			hologram.appendTextLine("§7Playing: §a" + this.playing/2);
-		} else if (this.type.getGameType().equals(GameType.DUEL)) {
-			double y = npc.getStoredLocation().getBlockY()+3;
-			Location l = new Location (npc.getStoredLocation().getWorld(),x,y,z);
-			this.hologram = HologramsAPI.createHologram(Main.get(),l);
-			hologram.clearLines();
+		}  else if (this.type.getGameType().equals(GameType.DUEL)) {
 			hologram.appendTextLine("§7Playing: §a" + this.playing);
 			hologram.appendTextLine("§7In Queue: §a" + this.queue);
 		}
 		}
 	}
+	
 	
 	public void updateHologram() {	
 		if (this.type.equals(NPCType.VOTES)) return;
@@ -54,13 +58,23 @@ public class SplinduxNPC {
 		} else {		
 			if (this.queue!= GameManager.getManager().getQueueSize(type.getSpleefType(), type.getGameType(), type.getAmount()) || 
 					this.playing!=GameManager.getManager().getPlayingSize(type.getSpleefType(), type.getGameType(), type.getAmount())) {
-				this.playing = GameManager.getManager().getPlayingSize(type.getSpleefType(), type.getGameType(), 0);
-				this.queue = GameManager.getManager().getQueueSize(type.getSpleefType(), type.getGameType(), 0);				
+				this.playing = GameManager.getManager().getPlayingSize(type.getSpleefType(), type.getGameType(), type.getAmount());
+				this.queue = GameManager.getManager().getQueueSize(type.getSpleefType(), type.getGameType(), type.getAmount());				
 				hologram.clearLines();
 				hologram.appendTextLine("§7Playing: §a" + this.playing);
 				hologram.appendTextLine("§7In Queue: §a" + this.queue);
 			}
 		}
+	}
+	
+	public Hologram loadHologram(Location l) {
+		for (Hologram h : HologramsAPI.getHolograms(Main.get())) {
+			if (h.getLocation().distance(l)<1) {
+				return h;
+			}
+		}
+		
+		return HologramsAPI.createHologram(Main.get(),l);
 	}
 	
 	public NPCType getType() {
@@ -72,7 +86,9 @@ public class SplinduxNPC {
 	}
 	
 	public void deleteHologram() {
+		if (!this.type.equals(NPCType.VOTES)) {
 		this.hologram.delete();
+		}
 	}
